Adds functional tests
* Use a heat template copied from sf to spawn sf and koji node
* Set local and instances /etc/hosts so that DNS works fine
* Configure koji using preprod inventory and extra_var settings

+++ koji/ansible/roles/mirror-rpmfactory/files/mirror-cloud.sh
@@ -8,7 +8,10 @@ if [ -f /var/lock/subsys/rsync_updates ]; then
 fi
 if [ -d ${TARGET} ] ; then
   touch /var/lock/subsys/rsync_updates
-  rsync  -avSHP --delete --exclude "local*" --exclude "isos" --exclude "instance" rsync://mirrors.ircam.fr/pub/CentOS/7/cloud/ ${TARGET}/
+  if [ "$(hostname -f)" != "koji.rpmfactory.sftests.com" ]; then
+    # Don't sync in pre-prod
+    rsync  -avSHP --delete --exclude "local*" --exclude "isos" --exclude "instance" rsync://mirrors.ircam.fr/pub/CentOS/7/cloud/ ${TARGET}/
+  fi
   /bin/rm -f /var/lock/subsys/rsync_updates
   echo "Change SELinux context, (httpd_sys_content_t)."
   chcon -Rv --type=httpd_sys_content_t ${TARGET}
